+++ spec/rubocop/cli_spec.rb
@@ -834,21 +834,18 @@ describe RuboCop::CLI, :isolated_environment do
 
       it 'does not say [Corrected] if correction was avoided' do
         src = ['# encoding: utf-8',
-               'a = c and b',
                'not a && b',
                'func a do b end',
                "Signal.trap('TERM') { system(cmd); exit }"]
         corrected = ['# encoding: utf-8',
-                     'a = c and b',
                      'not a && b',
                      'func a do b end',
                      "Signal.trap('TERM') { system(cmd); exit }"]
         offenses =
           ['== example.rb ==',
-           'C:  2:  7: Use && instead of and.',
-           'C:  3:  1: Use ! instead of not.',
-           'C:  4:  8: Prefer {...} over do...end for single-line blocks.',
-           'C:  5: 34: Do not use semicolons to terminate expressions.']
+           'C:  2:  1: Use ! instead of not.',
+           'C:  3:  8: Prefer {...} over do...end for single-line blocks.',
+           'C:  4: 34: Do not use semicolons to terminate expressions.']
 
         if RUBY_VERSION >= '2'
           src += ['def self.some_method(foo, bar: 1)',
@@ -857,10 +854,10 @@ describe RuboCop::CLI, :isolated_environment do
           corrected += ['def self.some_method(foo, bar: 1)',
                         '  log.debug(foo)',
                         'end']
-          offenses += ['W:  6: 27: Unused method argument - bar.']
-          summary = '1 file inspected, 5 offenses detected'
-        else
+          offenses += ['W:  5: 27: Unused method argument - bar.']
           summary = '1 file inspected, 4 offenses detected'
+        else
+          summary = '1 file inspected, 3 offenses detected'
         end
         create_file('example.rb', src)
         expect(cli.run(%w(-a -f simple))).to eq(1)